requirements.txt
PyQt6==6.6.1
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.8.2
pyqtgraph==0.13.3
pillow==10.1.0